Simple addition calculator
Used to learn MVC concepts, takes in 2 numbers and adds them.  Uses a Model, View and Controller class.  The Model class takes the 2 numbers entered in the View, by a user and then calculates and stores the solution.  The View class is all of the components of the "calculator".  Button, Labels, etc.  The Controller class works with the Model and View class to pass the inputs from the View to the Model, where the calculation is performed and then return from the Model back to the View, the solution.